requirements.txt
Django==1.10
djangorestframework==3.4.4
pep8==1.7.0
